requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
attrs
blinker
certifi
charset-normalizer
click
colorama
distro
Flask
frozenlist
h11
httpcore
httpx
idna
itsdangerous
Jinja2
jiter
joblib
MarkupSafe
multidict
numpy
openai
pandas
propcache
pydantic
pydantic_core
python-dateutil
python-dotenv
pytz
requests
scikit-learn
scipy
six
sniffio
threadpoolctl
tqdm
typing_extensions
tzdata
urllib3
Werkzeug
yarl
sqlalchemy 
psycopg2-binary
 